[improve](cloud) Txn lazy committer commit partitions in parallel
