+++ cloud/src/meta-service/txn_lazy_committer.cpp
@@ -29,6 +29,7 @@
 #include "common/defer.h"
 #include "common/logging.h"
 #include "common/stats.h"
+#include "common/sync_executor.h"
 #include "common/util.h"
 #include "cpp/sync_point.h"
 #include "meta-service/meta_service_helper.h"
@@ -633,23 +634,12 @@ void TxnLazyCommitTask::commit() {
             // <partition_id, tmp_rowsets>
             std::map<int64_t, std::vector<std::pair<std::string, doris::RowsetMetaCloudPB>>>
                     partition_to_tmp_rowset_metas;
-            size_t max_rowset_meta_size = 0;
             for (auto& [tmp_rowset_key, tmp_rowset_pb] : all_tmp_rowset_metas) {
                 partition_to_tmp_rowset_metas[tmp_rowset_pb.partition_id()].emplace_back();
                 partition_to_tmp_rowset_metas[tmp_rowset_pb.partition_id()].back().first =
                         tmp_rowset_key;
                 partition_to_tmp_rowset_metas[tmp_rowset_pb.partition_id()].back().second =
                         tmp_rowset_pb;
-                max_rowset_meta_size = std::max(max_rowset_meta_size, tmp_rowset_pb.ByteSizeLong());
-            }
-
-            // fdb txn limit 10MB, we use 4MB as the max size for each batch.
-            size_t max_rowsets_per_batch = config::txn_lazy_max_rowsets_per_batch;
-            if (max_rowset_meta_size > 0) {
-                max_rowsets_per_batch = std::min((4UL << 20) / max_rowset_meta_size,
-                                                 size_t(config::txn_lazy_max_rowsets_per_batch));
-                TEST_SYNC_POINT_CALLBACK("TxnLazyCommitTask::commit::max_rowsets_per_batch",
-                                         &max_rowsets_per_batch, &max_rowset_meta_size);
             }
 
             // Shuffle partition ids to reduce the conflict probability
@@ -662,134 +652,40 @@ void TxnLazyCommitTask::commit() {
                 std::shuffle(partition_ids.begin(), partition_ids.end(), rng);
             }
 
-            for (int64_t partition_id : partition_ids) {
-                auto& tmp_rowset_metas = partition_to_tmp_rowset_metas[partition_id];
-
-                // tablet_id -> TabletIndexPB
-                std::map<int64_t, TabletIndexPB> tablet_ids;
-                int64_t table_id = -1;
-                {
-                    DCHECK(tmp_rowset_metas.size() > 0);
-                    int64_t first_tablet_id = tmp_rowset_metas.begin()->second.tablet_id();
-                    TabletIndexPB first_tablet_index;
-                    std::tie(code_, msg_) = get_tablet_index(
-                            meta_reader, txn_kv_.get(), instance_id_, txn_id_, first_tablet_id,
-                            &first_tablet_index, is_versioned_read);
-                    if (code_ != MetaServiceCode::OK) {
+            if (config::enable_cloud_parallel_txn_lazy_commit) {
+                SyncExecutor<std::pair<MetaServiceCode, std::string>> executor(
+                        txn_lazy_committer_->parallel_commit_pool(),
+                        fmt::format("txn_{}_parallel_commit", txn_id_));
+                for (int64_t partition_id : partition_ids) {
+                    executor.add([&, partition_id, this]() {
+                        return commit_partition(db_id, partition_id,
+                                                partition_to_tmp_rowset_metas.at(partition_id),
+                                                is_versioned_read, is_versioned_write);
+                    });
+                }
+                bool finished = false;
+                auto task_results = executor.when_all(&finished);
+                for (auto&& [code, msg] : task_results) {
+                    if (code != MetaServiceCode::OK) {
+                        code_ = code;
+                        msg_ = std::move(msg);
                         break;
                     }
-                    LOG(INFO) << "first_tablet_index: " << first_tablet_index.ShortDebugString();
-                    table_id = first_tablet_index.table_id();
-                    tablet_ids.emplace(first_tablet_id, first_tablet_index);
                 }
-
-                // The partition version key is constructed during the txn commit process, so the versionstamp
-                // can be retrieved from the key itself.
-                Versionstamp versionstamp;
-                VersionPB original_partition_version;
-                std::tie(code_, msg_) = get_partition_version(
-                        txn_kv_.get(), instance_id_, txn_id_, db_id, table_id, partition_id,
-                        &original_partition_version, &versionstamp, is_versioned_read);
-                if (code_ != MetaServiceCode::OK) {
+                if (code_ == MetaServiceCode::OK && !finished) {
+                    LOG(WARNING) << "some partition commit tasks not finished, txn_id=" << txn_id_;
+                    code_ = MetaServiceCode::KV_TXN_CONFLICT;
                     break;
-                } else if (original_partition_version.pending_txn_ids_size() == 0 ||
-                           original_partition_version.pending_txn_ids(0) != txn_id_) {
-                    // The partition version does not contain the target pending txn, it might have been committed
-                    LOG(INFO) << "txn_id=" << txn_id_ << " partition_id=" << partition_id
-                              << " version has already been converted."
-                              << " version_pb:" << original_partition_version.ShortDebugString();
-                    TEST_SYNC_POINT_CALLBACK("TxnLazyCommitTask::commit::already_been_converted",
-                                             &original_partition_version);
-                    continue;
                 }
-
-                for (size_t i = 0; i < tmp_rowset_metas.size(); i += max_rowsets_per_batch) {
-                    size_t end = (i + max_rowsets_per_batch) > tmp_rowset_metas.size()
-                                         ? tmp_rowset_metas.size()
-                                         : i + max_rowsets_per_batch;
-                    std::vector<std::pair<std::string, doris::RowsetMetaCloudPB>>
-                            sub_partition_tmp_rowset_metas(tmp_rowset_metas.begin() + i,
-                                                           tmp_rowset_metas.begin() + end);
-                    convert_tmp_rowsets(instance_id_, txn_id_, txn_kv_, code_, msg_, db_id,
-                                        sub_partition_tmp_rowset_metas, tablet_ids,
-                                        is_versioned_write, is_versioned_read, versionstamp,
-                                        txn_lazy_committer_->resource_manager().get());
+            } else {
+                for (int64_t partition_id : partition_ids) {
+                    std::tie(code_, msg_) = commit_partition(
+                            db_id, partition_id, partition_to_tmp_rowset_metas[partition_id],
+                            is_versioned_read, is_versioned_write);
                     if (code_ != MetaServiceCode::OK) break;
                 }
-                if (code_ != MetaServiceCode::OK) break;
-
-                std::unique_ptr<Transaction> txn;
-                TxnErrorCode err = txn_kv_->create_txn(&txn);
-                if (err != TxnErrorCode::TXN_OK) {
-                    code_ = cast_as<ErrCategory::CREATE>(err);
-                    ss << "failed to create txn, txn_id=" << txn_id_ << " err=" << err;
-                    msg_ = ss.str();
-                    LOG(WARNING) << msg_;
-                    break;
-                }
-
-                DCHECK(table_id > 0);
-                DCHECK(partition_id > 0);
-
-                // Notice: get the versionstamp again, since it must not be changed here.
-                // But if you want to support multi pending txns in the partition version, this must be changed.
-                VersionPB version_pb;
-                std::tie(code_, msg_) = get_partition_version(
-                        txn.get(), instance_id_, txn_id_, db_id, table_id, partition_id,
-                        &version_pb, &versionstamp, is_versioned_read);
-                if (code_ != MetaServiceCode::OK) {
-                    break;
-                }
-
-                if (version_pb.pending_txn_ids_size() > 0 &&
-                    version_pb.pending_txn_ids(0) == txn_id_) {
-                    DCHECK(version_pb.pending_txn_ids_size() == 1);
-                    version_pb.clear_pending_txn_ids();
-
-                    if (version_pb.has_version()) {
-                        version_pb.set_version(version_pb.version() + 1);
-                    } else {
-                        // first commit txn version is 2
-                        version_pb.set_version(2);
-                    }
-                    std::string ver_key =
-                            partition_version_key({instance_id_, db_id, table_id, partition_id});
-                    std::string ver_val;
-                    if (!version_pb.SerializeToString(&ver_val)) {
-                        code_ = MetaServiceCode::PROTOBUF_SERIALIZE_ERR;
-                        ss << "failed to serialize version_pb when saving, txn_id=" << txn_id_;
-                        msg_ = ss.str();
-                        break;
-                    }
-                    txn->put(ver_key, ver_val);
-                    LOG(INFO) << "put ver_key=" << hex(ver_key) << " txn_id=" << txn_id_
-                              << " version_pb=" << version_pb.ShortDebugString();
-                    if (is_versioned_write) {
-                        // Update the partition version with the specified versionstamp.
-                        std::string versioned_key =
-                                versioned::partition_version_key({instance_id_, partition_id});
-                        versioned_put(txn.get(), versioned_key, versionstamp, ver_val);
-                        LOG(INFO) << "put versioned ver_key=" << hex(versioned_key)
-                                  << " txn_id=" << txn_id_
-                                  << " version_pb=" << version_pb.ShortDebugString();
-                    }
-
-                    for (auto& [tmp_rowset_key, tmp_rowset_pb] : tmp_rowset_metas) {
-                        txn->remove(tmp_rowset_key);
-                        LOG(INFO) << "remove tmp_rowset_key=" << hex(tmp_rowset_key)
-                                  << " txn_id=" << txn_id_;
-                    }
-
-                    TEST_SYNC_POINT_CALLBACK("TxnLazyCommitter::commit");
-                    err = txn->commit();
-                    if (err != TxnErrorCode::TXN_OK) {
-                        code_ = cast_as<ErrCategory::COMMIT>(err);
-                        ss << "failed to commit kv txn, txn_id=" << txn_id_ << " err=" << err;
-                        msg_ = ss.str();
-                        break;
-                    }
-                }
             }
+
             if (code_ != MetaServiceCode::OK) {
                 LOG(WARNING) << "txn_id=" << txn_id_ << " code=" << code_ << " msg=" << msg_;
                 break;
@@ -798,6 +694,158 @@ void TxnLazyCommitTask::commit() {
         } while (false);
     } while (code_ == MetaServiceCode::KV_TXN_CONFLICT &&
              retry_times++ < config::txn_store_retry_times);
+}
+
+std::pair<MetaServiceCode, std::string> TxnLazyCommitTask::commit_partition(
+        int64_t db_id, int64_t partition_id,
+        const std::vector<std::pair<std::string, doris::RowsetMetaCloudPB>>& tmp_rowset_metas,
+        bool is_versioned_read, bool is_versioned_write) {
+    std::stringstream ss;
+    CloneChainReader meta_reader(instance_id_, txn_kv_.get(),
+                                 txn_lazy_committer_->resource_manager().get());
+    MetaServiceCode code;
+    std::string msg;
+
+    StopWatch sw;
+    DORIS_CLOUD_DEFER {
+        g_bvar_txn_lazy_committer_commit_partition_duration << sw.elapsed_us();
+    };
+
+    // tablet_id -> TabletIndexPB
+    std::map<int64_t, TabletIndexPB> tablet_ids;
+    int64_t table_id = -1;
+    {
+        DCHECK(tmp_rowset_metas.size() > 0);
+        int64_t first_tablet_id = tmp_rowset_metas.begin()->second.tablet_id();
+        TabletIndexPB first_tablet_index;
+        std::tie(code, msg) =
+                get_tablet_index(meta_reader, txn_kv_.get(), instance_id_, txn_id_, first_tablet_id,
+                                 &first_tablet_index, is_versioned_read);
+        if (code != MetaServiceCode::OK) {
+            return {code, msg};
+        }
+        table_id = first_tablet_index.table_id();
+        tablet_ids.emplace(first_tablet_id, first_tablet_index);
+    }
+
+    // The partition version key is constructed during the txn commit process, so the versionstamp
+    // can be retrieved from the key itself.
+    Versionstamp versionstamp;
+    VersionPB original_partition_version;
+    std::tie(code, msg) = get_partition_version(txn_kv_.get(), instance_id_, txn_id_, db_id,
+                                                table_id, partition_id, &original_partition_version,
+                                                &versionstamp, is_versioned_read);
+    if (code != MetaServiceCode::OK) {
+        return {code, msg};
+    } else if (original_partition_version.pending_txn_ids_size() == 0 ||
+               original_partition_version.pending_txn_ids(0) != txn_id_) {
+        // The partition version does not contain the target pending txn, it might have been committed
+        LOG(INFO) << "txn_id=" << txn_id_ << " partition_id=" << partition_id
+                  << " version has already been converted."
+                  << " version_pb:" << original_partition_version.ShortDebugString();
+        TEST_SYNC_POINT_CALLBACK("TxnLazyCommitTask::commit::already_been_converted",
+                                 &original_partition_version);
+        return {MetaServiceCode::OK, ""};
+    }
+
+    size_t max_rowset_meta_size = 0;
+    for (const auto& [_, tmp_rowset_pb] : tmp_rowset_metas) {
+        max_rowset_meta_size = std::max(max_rowset_meta_size, tmp_rowset_pb.ByteSizeLong());
+    }
+
+    // fdb txn limit 10MB, we use 4MB as the max size for each batch.
+    size_t max_rowsets_per_batch = config::txn_lazy_max_rowsets_per_batch;
+    if (max_rowset_meta_size > 0) {
+        max_rowsets_per_batch = std::min((4UL << 20) / max_rowset_meta_size,
+                                         size_t(config::txn_lazy_max_rowsets_per_batch));
+        TEST_SYNC_POINT_CALLBACK("TxnLazyCommitTask::commit::max_rowsets_per_batch",
+                                 &max_rowsets_per_batch, &max_rowset_meta_size);
+    }
+
+    for (size_t i = 0; i < tmp_rowset_metas.size(); i += max_rowsets_per_batch) {
+        size_t end = (i + max_rowsets_per_batch) > tmp_rowset_metas.size()
+                             ? tmp_rowset_metas.size()
+                             : i + max_rowsets_per_batch;
+        std::vector<std::pair<std::string, doris::RowsetMetaCloudPB>>
+                sub_partition_tmp_rowset_metas(tmp_rowset_metas.begin() + i,
+                                               tmp_rowset_metas.begin() + end);
+        convert_tmp_rowsets(instance_id_, txn_id_, txn_kv_, code, msg, db_id,
+                            sub_partition_tmp_rowset_metas, tablet_ids, is_versioned_write,
+                            is_versioned_read, versionstamp,
+                            txn_lazy_committer_->resource_manager().get());
+        if (code != MetaServiceCode::OK) {
+            return {code, msg};
+        }
+    }
+
+    std::unique_ptr<Transaction> txn;
+    TxnErrorCode err = txn_kv_->create_txn(&txn);
+    if (err != TxnErrorCode::TXN_OK) {
+        code = cast_as<ErrCategory::CREATE>(err);
+        ss << "failed to create txn, txn_id=" << txn_id_ << " err=" << err;
+        msg = ss.str();
+        LOG(WARNING) << msg;
+        return {code, msg};
+    }
+
+    DCHECK(table_id > 0);
+    DCHECK(partition_id > 0);
+
+    // Notice: get the versionstamp again, since it must not be changed here.
+    // But if you want to support multi pending txns in the partition version, this must be changed.
+    VersionPB version_pb;
+    std::tie(code, msg) =
+            get_partition_version(txn.get(), instance_id_, txn_id_, db_id, table_id, partition_id,
+                                  &version_pb, &versionstamp, is_versioned_read);
+    if (code != MetaServiceCode::OK) {
+        return {code, msg};
+    }
+
+    if (version_pb.pending_txn_ids_size() > 0 && version_pb.pending_txn_ids(0) == txn_id_) {
+        DCHECK(version_pb.pending_txn_ids_size() == 1);
+        version_pb.clear_pending_txn_ids();
+
+        if (version_pb.has_version()) {
+            version_pb.set_version(version_pb.version() + 1);
+        } else {
+            // first commit txn version is 2
+            version_pb.set_version(2);
+        }
+        std::string ver_key = partition_version_key({instance_id_, db_id, table_id, partition_id});
+        std::string ver_val;
+        if (!version_pb.SerializeToString(&ver_val)) {
+            code = MetaServiceCode::PROTOBUF_SERIALIZE_ERR;
+            ss << "failed to serialize version_pb when saving, txn_id=" << txn_id_;
+            msg = ss.str();
+            return {code, msg};
+        }
+        txn->put(ver_key, ver_val);
+        LOG(INFO) << "put ver_key=" << hex(ver_key) << " txn_id=" << txn_id_
+                  << " version_pb=" << version_pb.ShortDebugString();
+        if (is_versioned_write) {
+            // Update the partition version with the specified versionstamp.
+            std::string versioned_key =
+                    versioned::partition_version_key({instance_id_, partition_id});
+            versioned_put(txn.get(), versioned_key, versionstamp, ver_val);
+            LOG(INFO) << "put versioned ver_key=" << hex(versioned_key) << " txn_id=" << txn_id_
+                      << " version_pb=" << version_pb.ShortDebugString();
+        }
+
+        for (auto& [tmp_rowset_key, tmp_rowset_pb] : tmp_rowset_metas) {
+            txn->remove(tmp_rowset_key);
+            LOG(INFO) << "remove tmp_rowset_key=" << hex(tmp_rowset_key) << " txn_id=" << txn_id_;
+        }
+
+        TEST_SYNC_POINT_CALLBACK("TxnLazyCommitter::commit");
+        err = txn->commit();
+        if (err != TxnErrorCode::TXN_OK) {
+            code = cast_as<ErrCategory::COMMIT>(err);
+            ss << "failed to commit kv txn, txn_id=" << txn_id_ << " err=" << err;
+            msg = ss.str();
+            return {code, msg};
+        }
+    }
+    return {MetaServiceCode::OK, ""};
 }
 
 std::pair<MetaServiceCode, std::string> TxnLazyCommitTask::wait() {
@@ -833,6 +881,14 @@ TxnLazyCommitter::TxnLazyCommitter(std::shared_ptr<TxnKv> txn_kv,
     worker_pool_ = std::make_unique<SimpleThreadPool>(config::txn_lazy_commit_num_threads,
                                                       "txn_lazy_commiter");
     worker_pool_->start();
+
+    int32_t parallel_commit_threads = std::max(0, config::parallel_txn_lazy_commit_num_threads);
+    if (parallel_commit_threads == 0) {
+        parallel_commit_threads = std::thread::hardware_concurrency();
+    }
+    parallel_commit_pool_ =
+            std::make_shared<SimpleThreadPool>(parallel_commit_threads, "parallel_commit");
+    parallel_commit_pool_->start();
 }
 
 /**
