+++ cloud/src/meta-service/txn_lazy_committer.h
@@ -21,7 +21,10 @@
 #include <bthread/mutex.h>
 #include <gen_cpp/cloud.pb.h>
 
+#include <memory>
+
 #include "common/simple_thread_pool.h"
+#include "meta-store/clone_chain_reader.h"
 #include "meta-store/txn_kv.h"
 #include "resource-manager/resource_manager.h"
 
@@ -42,6 +45,11 @@ public:
 private:
     friend class TxnLazyCommitter;
 
+    std::pair<MetaServiceCode, std::string> commit_partition(
+            int64_t db_id, int64_t partition_id,
+            const std::vector<std::pair<std::string, doris::RowsetMetaCloudPB>>& tmp_rowset_metas,
+            bool is_versioned_write, bool is_versioned_read);
+
     std::string instance_id_;
     int64_t txn_id_;
     std::shared_ptr<TxnKv> txn_kv_;
@@ -61,12 +69,14 @@ public:
     void remove(int64_t txn_id);
 
     std::shared_ptr<ResourceManager>& resource_manager() { return resource_mgr_; }
+    std::shared_ptr<SimpleThreadPool>& parallel_commit_pool() { return parallel_commit_pool_; }
 
 private:
     std::shared_ptr<TxnKv> txn_kv_;
     std::shared_ptr<ResourceManager> resource_mgr_;
 
     std::unique_ptr<SimpleThreadPool> worker_pool_;
+    std::shared_ptr<SimpleThreadPool> parallel_commit_pool_;
 
     std::mutex mutex_;
     // <txn_id, TxnLazyCommitTask>
